requirements.txt
aiogram==3.22.0
requests==2.32.5
python-dotenv==1.1.1